+++ lib/factory.ts
@@ -10,7 +10,6 @@ import isIp from './is-ip';
 import isValidHostname from './is-valid';
 import { IPublicSuffix, ISuffixLookupOptions } from './lookup/interface';
 import { IOptions, setDefaults } from './options';
-import getPublicSuffix from './public-suffix';
 import getSubdomain from './subdomain';
 
 export interface IResult {
@@ -21,7 +20,7 @@ export interface IResult {
   hostname: string | null;
 
   // Is `hostname` an IP? (IPv4 or IPv6)
-  isIp: boolean;
+  isIp: boolean | null;
 
   // `hostname` split between subdomain, domain and its public suffix (if any)
   subdomain: string | null;
@@ -49,7 +48,11 @@ export const enum FLAG {
 export default function parseImpl(
   url: string,
   step: FLAG,
-  suffixLookup: (_1: string, _2: ISuffixLookupOptions) => IPublicSuffix | null,
+  suffixLookup: (
+    _1: string,
+    _2: ISuffixLookupOptions,
+    _3: IPublicSuffix,
+  ) => void,
   partialOptions?: Partial<IOptions>,
 ): IResult {
   const options: IOptions = setDefaults(partialOptions);
@@ -57,52 +60,62 @@ export default function parseImpl(
     domain: null,
     hostname: null,
     isIcann: null,
-    isIp: false,
+    isIp: null,
     isPrivate: null,
     publicSuffix: null,
     subdomain: null,
   };
 
-  // Extract hostname from `url` only if needed
+  // Very fast approximate check to make sure `url` is a string. This is needed
+  // because the library will not necessarily be used in a typed setup and
+  // values of arbitrary types might be given as argument.
+  if (typeof url !== 'string') {
+    return result;
+  }
+
+  // Extract hostname from `url` only if needed. This can be made optional
+  // using `options.extractHostname`. This option will typically be used
+  // whenever we are sure the inputs to `parse` are already hostnames and not
+  // arbitrary URLs.
+  //
+  // `mixedInput` allows to specify if we expect a mix of URLs and hostnames
+  // as input. If only hostnames are expected then `extractHostname` can be
+  // set to `false` to speed-up parsing. If only URLs are expected then
+  // `mixedInputs` can be set to `false`. The `mixedInputs` is only a hint
+  // and will not change the behavior of the library.
   if (options.extractHostname === false) {
     result.hostname = url;
-
-    if (step === FLAG.HOSTNAME) {
-      return result;
-    }
+  } else if (options.mixedInputs === true) {
+    result.hostname = extractHostname(url, isValidHostname(url));
   } else {
-    const urlIsValidHostname = isValidHostname(url);
-    result.hostname = extractHostname(url, urlIsValidHostname);
+    result.hostname = extractHostname(url, false);
+  }
 
-    if (step === FLAG.HOSTNAME || result.hostname === null) {
-      return result;
-    }
+  if (step === FLAG.HOSTNAME || result.hostname === null) {
+    return result;
+  }
 
-    // Check if `hostname` is a valid ip address
+  // Check if `hostname` is a valid ip address
+  if (options.detectIp === true) {
     result.isIp = isIp(result.hostname);
-    if (result.isIp) {
-      return result;
-    }
-
-    // Make sure hostname is valid before proceeding
-    if (
-      urlIsValidHostname === false &&
-      isValidHostname(result.hostname) === false
-    ) {
+    if (result.isIp === true) {
       return result;
     }
   }
 
-  // Extract public suffix
-  const publicSuffixResult = getPublicSuffix(
-    result.hostname,
-    options,
-    suffixLookup,
-  );
+  // Perform optional hostname validation. If hostname is not valid, no need to
+  // go further as there will be no valid domain or sub-domain.
+  if (
+    options.validateHostname === true &&
+    options.extractHostname === true &&
+    isValidHostname(result.hostname) === false
+  ) {
+    result.hostname = null;
+    return result;
+  }
 
-  result.publicSuffix = publicSuffixResult.publicSuffix;
-  result.isIcann = publicSuffixResult.isIcann;
-  result.isPrivate = publicSuffixResult.isIcann === false;
+  // Extract public suffix
+  suffixLookup(result.hostname, options, result);
   if (step === FLAG.PUBLIC_SUFFIX || result.publicSuffix === null) {
     return result;
   }