+++ lib/lookup/suffix-trie.ts
@@ -1,4 +1,4 @@
-import { exceptions, rules } from './data/trie';
+import { exceptions, ITrie, rules } from './data/trie';
 import { IPublicSuffix, ISuffixLookupOptions } from './interface';
 
 // Flags used to know if a rule is ICANN or Private
@@ -10,6 +10,7 @@ const enum RULE_TYPE {
 interface IMatch {
   index: number;
   isIcann: boolean;
+  isPrivate: boolean;
 }
 
 /**
@@ -17,33 +18,35 @@ interface IMatch {
  */
 function lookupInTrie(
   parts: string[],
-  trie: any,
+  trie: ITrie,
   index: number,
   allowedMask: number,
-): IMatch {
-  let node = trie;
-  const lookupResult: IMatch = {
-    index: -1,
-    isIcann: false,
-  };
-
+): IMatch | null {
+  let result: IMatch | null = null;
+  let node: ITrie | undefined = trie;
   while (node !== undefined) {
     // We have a match!
-    if (node.$ !== undefined && (node.$ & allowedMask) !== 0) {
-      lookupResult.index = index + 1;
-      lookupResult.isIcann = node.$ === RULE_TYPE.ICANN;
+    if ((node.$ & allowedMask) !== 0) {
+      result = {
+        index: index + 1,
+        isIcann: node.$ === RULE_TYPE.ICANN,
+        isPrivate: node.$ === RULE_TYPE.PRIVATE,
+      };
     }
 
     // No more `parts` to look for
     if (index === -1) {
-      return lookupResult;
+      break;
     }
 
-    node = node[parts[index]] || node['*'];
+    const succ: {
+      [label: string]: ITrie;
+    } = node.succ;
+    node = succ && (succ[parts[index]] || succ['*']);
     index -= 1;
   }
 
-  return lookupResult;
+  return result;
 }
 
 /**
@@ -52,54 +55,119 @@ function lookupInTrie(
 export default function suffixLookup(
   hostname: string,
   options: ISuffixLookupOptions,
-): IPublicSuffix | null {
-  const allowIcannDomains = options.allowIcannDomains;
-  const allowPrivateDomains = options.allowPrivateDomains;
+  out: IPublicSuffix,
+): void {
+  // Fast path for very popular suffixes; this allows to by-pass lookup
+  // completely as well as any extra allocation or string manipulation.
+  if (options.allowPrivateDomains === false && hostname.length > 3) {
+    const last: number = hostname.length - 1;
+    const c3: number = hostname.charCodeAt(last);
+    const c2: number = hostname.charCodeAt(last - 1);
+    const c1: number = hostname.charCodeAt(last - 2);
+    const c0: number = hostname.charCodeAt(last - 3);
+
+    if (
+      c3 === 109 /* 'm' */ &&
+      c2 === 111 /* 'o' */ &&
+      c1 === 99 /* 'c' */ &&
+      c0 === 46 /* '.' */
+    ) {
+      out.isIcann = true;
+      out.isPrivate = false;
+      out.publicSuffix = 'com';
+      return;
+    } else if (
+      c3 === 103 /* 'g' */ &&
+      c2 === 114 /* 'r' */ &&
+      c1 === 111 /* 'o' */ &&
+      c0 === 46 /* '.' */
+    ) {
+      out.isIcann = true;
+      out.isPrivate = false;
+      out.publicSuffix = 'org';
+      return;
+    } else if (
+      c3 === 117 /* 'u' */ &&
+      c2 === 100 /* 'd' */ &&
+      c1 === 101 /* 'e' */ &&
+      c0 === 46 /* '.' */
+    ) {
+      out.isIcann = true;
+      out.isPrivate = false;
+      out.publicSuffix = 'edu';
+      return;
+    } else if (
+      c3 === 118 /* 'v' */ &&
+      c2 === 111 /* 'o' */ &&
+      c1 === 103 /* 'g' */ &&
+      c0 === 46 /* '.' */
+    ) {
+      out.isIcann = true;
+      out.isPrivate = false;
+      out.publicSuffix = 'gov';
+      return;
+    } else if (
+      c3 === 116 /* 't' */ &&
+      c2 === 101 /* 'e' */ &&
+      c1 === 110 /* 'n' */ &&
+      c0 === 46 /* '.' */
+    ) {
+      out.isIcann = true;
+      out.isPrivate = false;
+      out.publicSuffix = 'net';
+      return;
+    } else if (
+      c3 === 101 /* 'e' */ &&
+      c2 === 100 /* 'd' */ &&
+      c1 === 46 /* '.' */
+    ) {
+      out.isIcann = true;
+      out.isPrivate = false;
+      out.publicSuffix = 'de';
+      return;
+    }
+  }
+
   const hostnameParts = hostname.split('.');
 
-  let allowedMask = 0;
-
-  // Define set of accepted public suffix (ICANN, PRIVATE)
-  if (allowPrivateDomains === true) {
-    allowedMask |= RULE_TYPE.PRIVATE;
-  }
-  if (allowIcannDomains === true) {
-    allowedMask |= RULE_TYPE.ICANN;
-  }
-
-  // Look for a match in rules
-  const lookupResult = lookupInTrie(
-    hostnameParts,
-    rules,
-    hostnameParts.length - 1,
-    allowedMask,
-  );
-
-  if (lookupResult.index === -1) {
-    return null;
-  }
+  const allowedMask =
+    (options.allowPrivateDomains === true ? RULE_TYPE.PRIVATE : 0) |
+    (options.allowIcannDomains === true ? RULE_TYPE.ICANN : 0);
 
   // Look for exceptions
-  const exceptionLookupResult = lookupInTrie(
+  const exceptionMatch = lookupInTrie(
     hostnameParts,
     exceptions,
     hostnameParts.length - 1,
     allowedMask,
   );
 
-  if (exceptionLookupResult.index !== -1) {
-    return {
-      isIcann: exceptionLookupResult.isIcann,
-      isPrivate: !exceptionLookupResult.isIcann,
-      publicSuffix: hostnameParts
-        .slice(exceptionLookupResult.index + 1)
-        .join('.'),
-    };
+  if (exceptionMatch !== null) {
+    out.isIcann = exceptionMatch.isIcann;
+    out.isPrivate = exceptionMatch.isPrivate;
+    out.publicSuffix = hostnameParts.slice(exceptionMatch.index + 1).join('.');
+    return;
   }
 
-  return {
-    isIcann: lookupResult.isIcann,
-    isPrivate: !lookupResult.isIcann,
-    publicSuffix: hostnameParts.slice(lookupResult.index).join('.'),
-  };
+  // Look for a match in rules
+  const rulesMatch = lookupInTrie(
+    hostnameParts,
+    rules,
+    hostnameParts.length - 1,
+    allowedMask,
+  );
+
+  if (rulesMatch !== null) {
+    out.isIcann = rulesMatch.isIcann;
+    out.isPrivate = rulesMatch.isPrivate;
+    out.publicSuffix = hostnameParts.slice(rulesMatch.index).join('.');
+    return;
+  }
+
+  // No match found...
+  // Prevailing rule is '*' so we consider the top-level domain to be the
+  // public suffix of `hostname` (e.g.: 'example.org' => 'org').
+  out.isIcann = false;
+  out.isPrivate = false;
+  out.publicSuffix = hostnameParts[hostnameParts.length - 1];
 }